Main Menu
Added main menu screen, we still need to decide what will go in the options setting, or if we want to add any other kind of buttons to the main menu.

Also slight issue when hovering over button not sure how to resolve, the text like "start" turns black but not sure how to change that setting.